+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Splunk SDK for Java Changelog
 
+## Version 1.3.1
+
+* Removed deprecated functions from `DistributedConfigurations`.
+* Added `Index.upload` oneshot searches to allow oneshot parameters.
+* Fixed a resource leak on a file load function.
+
 ## Version 1.3
 
 ### New features and APIs